Revert "feat(useFormControl): Add new hooks 'useFormControl'"
