+++ src/App.tsx
@@ -75,12 +75,11 @@ const App = () => {
             <div className="form__item-field">
               <FieldControl name="text">
                 {(props: any) => {
-                  const { value, setValue, touched, errors, pristine, dirty } = props;
+                  const { value, setValue, touched, errors } = props;
                   return (
                     <>
                       <input type="text" value={value} onChange={e => setValue(e.target.value)} />
                       {touched && errors && JSON.stringify(errors)}
-                      {JSON.stringify({ pristine, dirty }, null, 2)}
                     </>
                   );
                 }}
