+++ src/lib/react-form-control/FieldControl.tsx
@@ -1,6 +1,7 @@
-import React from "react";
+import React, { useContext } from "react";
+
+import { FieldContext } from "./FormGroupContext";
 import { ValidatorErrors } from "./Validators";
-import { useFormControl } from "./useFormControl";
 
 export interface FieldControlChildrenProps {
   value: any;
@@ -19,11 +20,19 @@ interface FieldControlProps {
 
 export const FieldControl: React.FC<FieldControlProps> = (props: FieldControlProps) => {
   const { name, children } = props;
-  const { value, errors, setValue, ...meta } = useFormControl(name);
+  const formGroup = useContext(FieldContext);
+  if (formGroup === null) {
+    throw new Error(
+      ['Could not find "formGroup" in context.', "Wrap the root component in a <FormGroupProvider>."].join(" ")
+    );
+  }
+
+  const { values, metaInfos, errors } = formGroup;
+  const setValue = (value: any) => formGroup.setValue({ [name]: value });
   return children({
-    value,
-    errors,
+    value: values[name],
+    ...metaInfos[name],
+    errors: errors ? errors[name] : null,
     setValue,
-    ...meta,
   });
 };
